requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.25.0
Django==3.2.25
django-cloudinary-storage==0.3.0
djangorestframework==3.15.1
idna==3.6
pillow==10.3.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
